Use Country (not Gender) if we're _in_south:
ChoiceField sets its own `choices` property to deal with South having removed
it, but this leads to an error (e.g. "'5' is not a valid choice") if a model
has a ChoiceField with more than 3 choices, and an instance with a
higher-numbered choice exists in the database during a migration.

This commit uses dj.choices.Country instead, which means we're now good up to
215 items...